Detector filtering

+++ include/detectors/BoldDetector.h
@@ -16,9 +16,10 @@ namespace visy {
 
         class BoldDetector : public Detector {
         public:
-            BoldDetector(std::vector<int>& sizes, int n_bins = 12, int scale_levels = 3, float scale_space_factor = 1.6f);
+            BoldDetector(std::vector<float>& sizes, int n_bins = 12, int scale_levels = 3, float scale_space_factor = 1.6f);
             virtual ~BoldDetector();
             virtual void detect(cv::Mat& source, pcl::PointCloud<PointType>::Ptr cloud, std::vector<visy::extractors::KeyPoint3D>& keypoints, cv::Mat& descriptor, cv::Mat* mask = NULL);
+            
             virtual std::string buildNameImpl();
         private:
             BoldLib::BOLD bold;
